small fixups

+++ src/utils/_backup/host.py
@@ -21,6 +21,9 @@ class BackupHost:
     mac: str = ""
     ping_retry: int = 20
     
+    _waked_up: bool = False
+    _power_off: bool = False
+
     @property
     def restic_target(self) -> str:
         hostname = socket.gethostname()
@@ -38,9 +41,9 @@ class BackupHost:
             return subprocess.call(cmd_splitted, stdout=output, stderr=output)  # noqa: S603
         except KeyboardInterrupt:
             sys.exit()
-            
+
     def wake_up(self) -> None:
-        if self.wol:
+        if self.wol and not self._waked_up:
             wol_cmd = f"wol {self.mac} -p 9"
             ping_cmd = f"ping -c 1 {self.host}"
             print(f"[BackupHost][wake_up][{self.host}] wol: send magic packet")
@@ -51,9 +54,11 @@ class BackupHost:
                 retry += 1
                 print(f"[BackupHost][wake_up][{self.host}] retry ping {retry}")
             print(f"[BackupHost][wake_up][{self.host}] got ping")
+            self._waked_up = True
 
     def poweroff(self) -> None:
-        if self.wol:
+        if self.wol and not self._power_off:
             poweroff_cmd = ["ssh", f"{self.user}@{self.host}", "sudo poweroff"]
             print(f"[BackupHost][poweroff][{self.host}] shutdown host")
             self._call(poweroff_cmd)
+            self._power_off = True
